Atualizando analisador e arquivos

+++ runner.sh
@@ -1,11 +1,15 @@
 
+clear
+
+if [ -d "./target" ]; then
+    rm -rf ./target
+fi
+
 mkdir ./target
 
-gcc -o ./target/analisador ./src/*.c ./algoritmos/*.c
+gcc -o ./target/analisador ./src/*.c ./algoritmos/*/*.c
 
-./target/analisador $1
+./target/analisador $@
 
 echo "Pressione qualquer tecla para continuar..."
 read -n 1 -s -r -p ""
-
-rm -rf ./target
